Fix: 전시목록 페이지 컴포넌트분리 #6
