+++ public/data/exhbBookmark.json
@@ -1,0 +1,1 @@
+"메이트글 북마크 등록이 완료되었습니다."
