+++ src/types/exhbList.d.ts
@@ -1,4 +1,4 @@
-export interface ExhbType {
+export interface Exhibition {
   id: number;
   title: string;
   posterUrl: string;
@@ -8,4 +8,17 @@ export interface ExhbType {
   status: string;
   space: string;
   viewCount: number;
+}
+
+export type StatusType = "현재" | "예정" | "지난";
+export type FilterType = "전체" | "영상" | "특별" | "기획" | "상설" | "소장품";
+
+export interface PagesState {
+  started: number;
+  selected: number;
+}
+
+export interface ExhbTypeFilters {
+  selected: FilterType;
+  applied: FilterType;
 }
